refactor: split Asset thunk into multiple files
